1、异常通用类名。 2、当前时间用整型。 3、JPA 事务不生效。

+++ server-magpie-sms/src/main/java/com/honeysense/magpie/sms/entity/SmsCode.java
@@ -41,7 +41,7 @@ public class SmsCode extends MagpieEntity {
     @NotNull
     @Min(20200101)
     @Max(20990101)
-    private Long day;
+    private Integer day;
     @ApiModelProperty(value="验证次数", example = "1", required = true)
     @NotNull
     @Min(0)
